+++ inst/apps/LowerBoiseBCGCalc/global.R
@@ -1,7 +1,7 @@
 # Shiny Global File
 
 # Version ----
-pkg_version <- "0.0.0.9013"
+pkg_version <- "0.0.0.9014"
 
 # Packages----
 library(BCGcalc)